invoice_kraken_ocr/pyproject.toml

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "invoice_kraken_ocr"
version = "0.1.0"
description = "Kraken OCR for handwritten Arabic invoices with incremental learning (ERPNext v15)"
readme = "invoice_kraken_ocr/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "kraken>=4.3.15",
  "Pillow>=10.0.0",
  "numpy",
  "opencv-python-headless",
  "rapidfuzz>=3.8.0"
]
